Refactor range calculation in trading strategies
Updated the `range_to_show` and `best_range_to_show` methods to refactor how price ranges are calculated in trading strategies. Removed redundant `range_to_show` and `best_range_to_show` methods from `CustomStrategy`, improving code maintainability and reducing duplication.
(ref. #22)
